avoid output height check during rendering.
